Move BreakInWorkHistoryComponent under candidate namespace
