+++ spec/components/candidate_interface/break_in_work_history_component_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe BreakInWorkHistoryComponent do
+RSpec.describe CandidateInterface::BreakInWorkHistoryComponent do
   let(:january2018) { Time.zone.local(2018, 1, 1) }
   let(:february2019) { Time.zone.local(2019, 2, 1) }
   let(:april2019) { Time.zone.local(2019, 4, 1) }
@@ -14,7 +14,7 @@ RSpec.describe BreakInWorkHistoryComponent do
   end
 
   it 'renders the component with the break length, reason and dates' do
-    result = render_inline(BreakInWorkHistoryComponent.new(work_break: work_break))
+    result = render_inline(CandidateInterface::BreakInWorkHistoryComponent.new(work_break: work_break))
 
     expect(result.text).to include('Break in work history (1 month)')
     expect(result.text).to include('I fell asleep.')
@@ -23,13 +23,13 @@ RSpec.describe BreakInWorkHistoryComponent do
 
   context 'when editable' do
     it 'renders the component with a delete link' do
-      result = render_inline(BreakInWorkHistoryComponent.new(work_break: work_break, editable: true))
+      result = render_inline(CandidateInterface::BreakInWorkHistoryComponent.new(work_break: work_break, editable: true))
 
       expect(result.text).to include('Delete entry for break between February 2019 and April 2019')
     end
 
     it 'renders the component with change links' do
-      result = render_inline(BreakInWorkHistoryComponent.new(work_break: work_break, editable: true))
+      result = render_inline(CandidateInterface::BreakInWorkHistoryComponent.new(work_break: work_break, editable: true))
 
       expect(result.text).to include('Change description for break between February 2019 and April 2019')
       expect(result.text).to include('Change dates for break between February 2019 and April 2019')
@@ -38,13 +38,13 @@ RSpec.describe BreakInWorkHistoryComponent do
 
   context 'when not editable' do
     it 'renders the component without a delete link' do
-      result = render_inline(BreakInWorkHistoryComponent.new(work_break: work_break, editable: false))
+      result = render_inline(CandidateInterface::BreakInWorkHistoryComponent.new(work_break: work_break, editable: false))
 
       expect(result.text).not_to include('Delete entry for break between February 2019 and April 2019')
     end
 
     it 'renders the component without change links' do
-      result = render_inline(BreakInWorkHistoryComponent.new(work_break: work_break, editable: false))
+      result = render_inline(CandidateInterface::BreakInWorkHistoryComponent.new(work_break: work_break, editable: false))
 
       expect(result.text).not_to include('Change description for break between February 2019 and April 2019')
       expect(result.text).not_to include('Change dates for break between February 2019 and April 2019')
